fix: imports to .js and .ts are resolved to the same file
The TypeScript compiler will recognize that an import to foo.js is actually an
import to foo.ts, if the .ts file exists and the .js file does not.

If two import paths are the same, ignoring a ".js" or ".ts" or ".jsx" suffix,
treat them as belonging to the same module.

Related to: https://github.com/stephenh/ts-proto/pull/602

Signed-off-by: Christian Stewart <christian@paral.in>